requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.36.17
botocore==1.36.17
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-js-asset==3.0.1
django-timezone-field==7.1
hiredis==3.1.0
idna==3.10
jmespath==1.0.1
kavenegar==1.1.2
kombu==5.4.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
